ViewModel-и перенесены в отдельный модуль
