+++ app/build.gradle
@@ -43,12 +43,12 @@ android {
 }
 
 dependencies {
-    implementation(project(":data"))
+    implementation(project(":viewModel"))
     implementation(project(":data:di"))
     implementation(project(":model"))
+    implementation(project(":core"))
 
     def fragment_version = "1.4.1"
-    def lifecycle_version = "2.4.1"
     def recycler_version = '1.2.1'
     def material_version = '1.5.0'
     def constraint_version = '2.1.3'
@@ -64,7 +64,6 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:$coroutines_version"
     implementation "androidx.fragment:fragment-ktx:$fragment_version"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_version"
     //noinspection LifecycleAnnotationProcessorWithJava8
