+++ app/src/main/java/ru/iteco/fmhandroid/adapter/ClaimListAdapter.kt
@@ -6,12 +6,9 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import ru.iteco.fmh.model.FullClaim
+import ru.iteco.fmh.viewmodel.claim.OnClaimItemClickListener
 import ru.iteco.fmhandroid.databinding.ItemClaimBinding
 import ru.iteco.fmhandroid.utils.Utils
-
-interface OnClaimItemClickListener {
-    fun onCard(fullClaim: FullClaim) {}
-}
 
 class ClaimListAdapter(
     private val onClaimItemClickListener: OnClaimItemClickListener
