+++ app/src/main/java/ru/iteco/fmhandroid/adapter/NewsListAdapter.kt
@@ -7,15 +7,12 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import ru.iteco.fmh.model.News
+import ru.iteco.fmh.viewmodel.news.NewsViewData
+import ru.iteco.fmh.viewmodel.news.OnNewsItemClickListener
 import ru.iteco.fmhandroid.R
 import ru.iteco.fmhandroid.databinding.ItemNewsBinding
 import ru.iteco.fmhandroid.extensions.getType
-import ru.iteco.fmhandroid.ui.viewdata.NewsViewData
 import ru.iteco.fmhandroid.utils.Utils
-
-interface OnNewsItemClickListener {
-    fun onCard(newsItem: NewsViewData)
-}
 
 class NewsListAdapter(private val onNewsItemClickListener: OnNewsItemClickListener) :
     ListAdapter<NewsViewData, NewsListAdapter.NewsViewHolder>(NewsDiffCallBack) {
