+++ app/src/main/java/ru/iteco/fmhandroid/ui/ClaimListFilteringDialogFragment.kt
@@ -14,7 +14,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import ru.iteco.fmh.model.Claim
 import ru.iteco.fmhandroid.R
-import ru.iteco.fmhandroid.viewmodel.ClaimViewModel
+import ru.iteco.fmh.viewmodel.claim.ClaimViewModel
 
 @AndroidEntryPoint
 class ClaimListFilteringDialogFragment : DialogFragment() {
