+++ app/src/main/java/ru/iteco/fmhandroid/ui/ClaimListFragment.kt
@@ -18,8 +18,8 @@ import ru.iteco.fmhandroid.R
 import ru.iteco.fmhandroid.adapter.ClaimListAdapter
 import ru.iteco.fmhandroid.databinding.FragmentListClaimBinding
 import ru.iteco.fmhandroid.dto.FullClaimNavArg
-import ru.iteco.fmhandroid.viewmodel.AuthViewModel
-import ru.iteco.fmhandroid.viewmodel.ClaimViewModel
+import ru.iteco.fmh.viewmodel.AuthViewModel
+import ru.iteco.fmh.viewmodel.claim.ClaimViewModel
 
 @AndroidEntryPoint
 class ClaimListFragment : Fragment(R.layout.fragment_list_claim) {
