+++ app/src/main/java/ru/iteco/fmhandroid/ui/MainFragment.kt
@@ -19,9 +19,9 @@ import ru.iteco.fmhandroid.adapter.NewsListAdapter
 import ru.iteco.fmhandroid.databinding.FragmentMainBinding
 import ru.iteco.fmhandroid.dto.FullClaimNavArg
 import ru.iteco.fmhandroid.utils.Utils
-import ru.iteco.fmhandroid.viewmodel.AuthViewModel
-import ru.iteco.fmhandroid.viewmodel.ClaimViewModel
-import ru.iteco.fmhandroid.viewmodel.NewsViewModel
+import ru.iteco.fmh.viewmodel.AuthViewModel
+import ru.iteco.fmh.viewmodel.claim.ClaimViewModel
+import ru.iteco.fmh.viewmodel.news.NewsViewModel
 
 @AndroidEntryPoint
 class MainFragment : Fragment(R.layout.fragment_main) {
