+++ app/src/main/java/ru/iteco/fmhandroid/ui/SplashScreenFragment.kt
@@ -13,12 +13,10 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
-import ru.iteco.fmh.data.AppAuth
 import ru.iteco.fmhandroid.R
 import ru.iteco.fmhandroid.databinding.FragmentSplashScreenBinding
 import ru.iteco.fmhandroid.dto.SplashScreenData
-import ru.iteco.fmhandroid.viewmodel.AuthViewModel
-import javax.inject.Inject
+import ru.iteco.fmh.viewmodel.AuthViewModel
 
 @SuppressLint("CustomSplashScreen")
 @AndroidEntryPoint
