--- app/src/main/java/ru/iteco/fmhandroid/ui/viewdata/OurMissionItemViewData.kt
@@ -1,13 +1,0 @@
-package ru.iteco.fmhandroid.ui.viewdata
-
-
-/**
- * Модель представления элемента раздела "Наша миссия" в списке
- */
-data class OurMissionItemViewData(
-    val id: Int,
-    val title: String,
-    val titleBackgroundColor: Int,
-    val description: String,
-    val isOpen: Boolean
-)
