+++ settings.gradle
@@ -2,3 +2,5 @@ rootProject.name = "FMHAndroid"
 include ':app'
 include ':model'
 include ':data:impl', ':data:di', ':data:db', ':data:api'
+include ':viewModel'
+include ':core'
