+++ viewModel/src/main/kotlin/ru/iteco/fmh/viewmodel/claim/OnClaimItemClickListener.kt
@@ -1,0 +1,7 @@
+package ru.iteco.fmh.viewmodel.claim
+
+import ru.iteco.fmh.model.FullClaim
+
+interface OnClaimItemClickListener {
+    fun onCard(fullClaim: FullClaim) {}
+}
