+++ viewModel/src/main/kotlin/ru/iteco/fmh/viewmodel/claim/card/OnClaimCommentItemClickListener.kt
@@ -1,0 +1,7 @@
+package ru.iteco.fmh.viewmodel.claim.card
+
+import ru.iteco.fmh.model.Claim
+
+interface OnClaimCommentItemClickListener {
+    fun onCard(claimComment: Claim.Comment)
+}
