+++ viewModel/src/main/kotlin/ru/iteco/fmh/viewmodel/news/OnNewsItemClickListener.kt
@@ -1,0 +1,7 @@
+package ru.iteco.fmh.viewmodel.news
+
+import ru.iteco.fmh.viewmodel.news.NewsViewData
+
+interface OnNewsItemClickListener {
+    fun onCard(newsItem: NewsViewData)
+}
